+++ tests/graphics/queue_ownership_test.c
@@ -162,7 +162,15 @@ PalBool queueOwnershipTest()
     }
 
     if (!cpyQueue) {
-        // we didnt get a copy queue that cannot share resource ownership with thr graphics queue
+        palLog(nullptr, "");
+        palLog(
+            nullptr, 
+            "Failed to find a copy queue that does not share ownership with the graphics queue");
+        
+        palLog(nullptr, "No ownership transfer is required.");
+        palLog(nullptr, "Defaulting to cross-queue synchronization.");
+        palLog(nullptr, "");
+
         result = palCreateQueue(device, PAL_QUEUE_TYPE_COPY, &cpyQueue);
         if (result != PAL_RESULT_SUCCESS) {
             logResult(result, "Failed to create queue");
@@ -260,7 +268,7 @@ PalBool queueOwnershipTest()
     palUnmapBuffer(stagingBuffer);
 
     // We use the copy queue to copy the buffer to the texture thus taking ownership
-    result = palCmdBegin(gfxCmdBuffer, nullptr);
+    result = palCmdBegin(cpyCmdBuffer, nullptr);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to begin command buffer");
         return PAL_FALSE;
@@ -278,30 +286,29 @@ PalBool queueOwnershipTest()
     barrierInfo.newState = PAL_USAGE_STATE_TRANSFER_WRITE;
     barrierInfo.dstStages = PAL_PIPELINE_STAGE_TRANSFER;
 
-    // palCmdImageBarrier(cpyCmdBuffer, texture, &range, &barrierInfo);
-    // palCmdCopyBufferToImage(cpyCmdBuffer, texture, stagingBuffer, &copyInfo);
+    palCmdImageBarrier(cpyCmdBuffer, texture, &range, &barrierInfo);
+    palCmdCopyBufferToImage(cpyCmdBuffer, texture, stagingBuffer, &copyInfo);
 
     // The copy queue is done with the texture, we transfer ownership to the graphics queue
-    PalBarrierInfo2 barrierInfo2 = {0};
-    barrierInfo2.oldState = PAL_USAGE_STATE_TRANSFER_WRITE;
-    barrierInfo2.srcStages = PAL_PIPELINE_STAGE_TRANSFER;
-    barrierInfo2.newState = PAL_USAGE_STATE_TRANSFER_READ;
-    barrierInfo2.dstStages = PAL_PIPELINE_STAGE_FRAGMENT_SHADER;
-
     // The graphics command buffer has to be in recording state.
-    // result = palCmdBegin(gfxCmdBuffer, nullptr);
-    // if (result != PAL_RESULT_SUCCESS) {
-    //     logResult(result, "Failed to begin command buffer");
-    //     return PAL_FALSE;
-    // }
+    result = palCmdBegin(gfxCmdBuffer, nullptr);
+    if (result != PAL_RESULT_SUCCESS) {
+        logResult(result, "Failed to begin command buffer");
+        return PAL_FALSE;
+    }
 
-    // If the two queues can share resource ownership, the function below just
-    // sets the barrier but does no ownership transfer
-    // barrierInfo2.dstCmdBuffer = gfxCmdBuffer; // The graphics queue command buffer
-    // palCmdImageBarrier2(cpyCmdBuffer, texture, &range, &barrierInfo2);
+    // this sets a normal barrier on the source command buffer is there is no
+    // ownership tranfer
+    palCmdImageOwnershipTransfer(
+        cpyCmdBuffer, 
+        gfxCmdBuffer, 
+        texture, 
+        &range, 
+        PAL_USAGE_STATE_TRANSFER_WRITE, 
+        PAL_PIPELINE_STAGE_TRANSFER);
 
     // End the copy command buffer
-    result = palCmdEnd(gfxCmdBuffer);
+    result = palCmdEnd(cpyCmdBuffer);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to end command buffer");
         return PAL_FALSE;
@@ -309,24 +316,27 @@ PalBool queueOwnershipTest()
 
     // Submit the copy command buffer
     PalCommandBufferSubmitInfo submitInfo = {0};
-    submitInfo.cmdBuffer = gfxCmdBuffer;
-    // submitInfo.fence = nullptr; // we are not using fence
-    // submitInfo.signalSemaphore = semaphore;
+    submitInfo.cmdBuffer = cpyCmdBuffer;
+    submitInfo.fence = nullptr; // we are not using fence
+    submitInfo.signalSemaphore = semaphore;
 
-    result = palSubmitCommandBuffer(gfxQueue, &submitInfo);
-    uint32_t lost = palGetDeviceLostReason(device);
+    result = palSubmitCommandBuffer(cpyQueue, &submitInfo);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to submit command buffer");
         return PAL_FALSE;
     }
 
-    // We transition the image to a different usage to check if the graphics queue has ownership
-    barrierInfo.oldState = PAL_USAGE_STATE_SHADER_READ;
-    barrierInfo.srcStages = PAL_PIPELINE_STAGE_FRAGMENT_SHADER;
-    barrierInfo.newState = PAL_USAGE_STATE_STORAGE_READ;
-    barrierInfo.dstStages = PAL_PIPELINE_STAGE_NONE;
-    // palCmdImageBarrier(gfxCmdBuffer, texture, &range, &barrierInfo);
+    // We transition the image to a different usage state after acquiring ownership
+    // After ownership transfer, the resource will always be in a default state
+    barrierInfo.oldState = PAL_USAGE_STATE_UNDEFINED;
+    barrierInfo.srcStages = PAL_PIPELINE_STAGE_NONE;
+    barrierInfo.newState = PAL_USAGE_STATE_SHADER_READ;
+    barrierInfo.dstStages = PAL_PIPELINE_STAGE_FRAGMENT_SHADER;
+    palCmdImageBarrier(gfxCmdBuffer, texture, &range, &barrierInfo);
 
+    // we could do additional work like reading back from the image and writing to a ppm file
+    // but we keep it very simple and focused
+    
     // End the graphics command buffer
     result = palCmdEnd(gfxCmdBuffer);
     if (result != PAL_RESULT_SUCCESS) {
@@ -338,6 +348,7 @@ PalBool queueOwnershipTest()
     submitInfo.cmdBuffer = gfxCmdBuffer;
     submitInfo.waitStages = PAL_PIPELINE_STAGE_TRANSFER;
     submitInfo.waitSemaphore = semaphore; // wait for the copy queue
+    submitInfo.fence = fence; // signal fence if graphics queue is done
 
     result = palSubmitCommandBuffer(gfxQueue, &submitInfo);
     if (result != PAL_RESULT_SUCCESS) {