+++ CHANGELOG.md
@@ -7,13 +7,12 @@
 
 ### Features
 
-- Added `palCmdImageBarrier2()` to transfer image ownership between two queues.
-- Added `palCmdBufferBarrier2()` to transfer buffer ownership between two queues.
+- Added `palCmdImageOwnershipTransfer()` to transfer image ownership between two queues.
+- Added `palCmdBufferOwnershipTransfer()` to transfer buffer ownership between two queues.
 - Added `palCanQueueShareOwnership()` to check if two queues can share resource ownership.
 - Added `palCanQueueUseUsageState()` to check if a queue can use a usage state.
 - Added `palCanQueueUsePipelineStages()` to check if a queue can use a pipeline stage(s).
 - Added `PAL_GRAPHICS_BACKEND_VTABLE_VERSION_2` constant to extend the graphics backend vtable.
-- Added struct `PalBarrierInfo2`.
 - Added struct `PalGraphicsBackendVtable2`.
 - Fixed D3D12 graphics backend queue bug where queues destroyed were not added back to the pool.
 - Fixed palFormatResult() bug where results without native code were not formatting properly.